requirements.txt
pandas
scikit-learn
flask
flask-cors
tensorflow
numpy
pymongo
python-dotenv
requests
vaderSentiment
